+++ frontend/src/components/DocumentPreview.js
@@ -11,7 +11,7 @@ import '../styles/index.css'
 import '../styles/documentpreview.css'
 
 import preview from '../media/preview.png'
-import WithAuth from './WithAuth'
+import WithAuth from './auth/WithAuth'
 
 // Not needed unless working with non "en" locales
 // addLocaleData(enLocaleData);
