+++ frontend/src/components/index.js
@@ -1,17 +1,20 @@
-import LogIn from './LogIn'
 import Dashboard from './Dashboard'
 import DocumentList from './DocumentList'
 import DocumentListItem from './DocumentListItem'
 import DocumentPreview from './DocumentPreview'
-import Register from './Register'
+
 import SelectDocumentsPage from './SelectDocuments'
 import BoxPreview from './DocumentPreview'
 import PMMainPage from './PMMainPage'
 import DocumentClassPage from './DocumentClassPage.js'
-import ForgotPassword from './ForgotPassword'
+
 import Load from './Load'
-import WrongPage from './WrongPage'
-import ErrorMessage from './ErrorMessage'
+
+import ErrorMessage from './auth/ErrorMessage'
+import ForgotPassword from './auth/ForgotPassword'
+import LogIn from './auth/LogIn'
+import Register from './auth/Register'
+import WrongPage from './auth/WrongPage'
 
 export {
   Dashboard,
